+++ crates/zizmor/tests/integration/audit/secrets_outside_env.rs
@@ -7,20 +7,21 @@ fn test_secrets_outside_env() -> Result<()> {
     insta::assert_snapshot!(
         zizmor()
             .input(input_under_test("secrets-outside-env.yml"))
+            .args(["--persona=auditor"])
             .run()?,
         @"
     warning[secrets-outside-env]: secrets referenced without a dedicated environment
-      --> @@INPUT@@:12:23
+      --> @@INPUT@@:20:20
        |
-     6 |   test:
+    12 |   test:
        |   ---- this job
     ...
-    12 |         run: echo ${{ secrets.FOO }}
-       |                       ^^^^^^^^^^^ secret is accessed outside of a dedicated environment
+    20 |           FOO: ${{ secrets.FOO }}
+       |                    ^^^^^^^^^^^ secret is accessed outside of a dedicated environment
        |
        = note: audit confidence → High
 
-    4 findings (3 suppressed): 0 informational, 0 low, 1 medium, 0 high
+    1 finding: 0 informational, 0 low, 1 medium, 0 high
     "
     );